Début de vérif Documentation de Entity (sauf contact) Documentation de Manager (sauf contact) Documentation de Controlleur (sauf contact) Doc de BackEnd Doc Vue/Autoloader/App

+++ Controlleur/ControlleurArticles.php
@@ -12,13 +12,18 @@ use Controlleur\Routeur\Routeur;
 use Modele\Entity\Articles;
 use Modele\Manager\ManagerArticles;
 use Vue\Core\Vue;
-
+/**
+ * Class ControlleurArticles utilisé pour la création de la page d'article
+ */
 class ControlleurArticles
 {
     private $articles;
     private $managerArt;
     private $traitement;
 
+    /**
+     * Constructeur pour instancier les outils CRUD
+     */
     public function __construct()
     {
         $this->articles = new Articles();
@@ -26,25 +31,36 @@ class ControlleurArticles
         $this->traitement = new Vue('traitement');
     }
 
+    /**
+     * Appeler par :  Routeur
+     * Méthode qui détecte l'action utilisé et redirige vers la méthode approprié
+     * @param $action
+     */
     public function traitement($action)
     {
-        if ($action==='create'){
-            $this->creerArticles();
-        }elseif ($action==='modif'){
-            $donnees = $this->managerArt->read($_GET['id']);
-            $this->majArticles($donnees);
-        }elseif ($action==='delete'){
-            $this->delete();
+        if (isset($action)){
+            if ($action==='create'){
+                $this->creerArticles();
+            }elseif ($action==='modif'){
+                $donnees = $this->managerArt->read($_GET['id']);
+                $this->majArticles($donnees);
+            }elseif ($action==='delete'){
+                $this->delete();
+            }
         }
     }
 
+    /**
+     * Méthode qui reçoit les données pour créé un article
+     * @return boolean
+    */
     public function creerArticles()
     {
         $this->traitement->genererPages();
         if ($_POST){
             if ($_POST['titre']===''){
                 return false;
-            }elseif( $_POST['contenu'] ===''){
+            }elseif($_POST['contenu'] ===''){
                 return false;
             }else{
                 $this->articles->setTitre($_POST['titre']);
@@ -56,6 +72,10 @@ class ControlleurArticles
         return true;
     }
 
+    /**
+     * Méthode qui reçoit les données pour modifier un article
+     * @return boolean
+     */
     public function majArticles($donnees)
     {
         $this->traitement->genererPages([$donnees]);
@@ -73,10 +93,14 @@ class ControlleurArticles
         return true;
     }
 
-
+    /**
+     * Méthode qui reçoit les données pour supprimer un article
+     * @return boolean
+     */
     public function delete()
     {
         $this->managerArt->delete($_GET['id']);
         Routeur::redirection('admin');
+        return true;
     }
 }
